+++ core/tests/unit/helpers/load_kmx_file.cpp
