+++ core/tests/unit/km_core_keyboard_api.tests.cpp
@@ -5,8 +5,8 @@
 #include <gtest/gtest.h>
 
 #include <keyman/keyman_core_api.h>
-#include "emscripten_filesystem.h"
-#include "load_kmx_file.hpp"
+#include "./helpers/emscripten_filesystem.h"
+#include "./helpers/load_kmx_file.hpp"
 
 km::core::path test_dir;
