+++ core/tests/unit/kmnkbd/action_api.tests.cpp
@@ -12,7 +12,7 @@
 #include "action.hpp"
 
 #include <test_assert.h>
-#include "../emscripten_filesystem.h"
+#include "../helpers/emscripten_filesystem.h"
 
 const km_core_action_item alert_action_item();
 const km_core_action_item bksp_action_item(uint8_t type, uintptr_t value);
