+++ core/tests/unit/kmnkbd/actions_get_api.tests.cpp
@@ -14,8 +14,8 @@
 #include "context.hpp"
 
 #include <test_assert.h>
-#include "../emscripten_filesystem.h"
-#include "../load_kmx_file.hpp"
+#include "../helpers/emscripten_filesystem.h"
+#include "../helpers/load_kmx_file.hpp"
 
 km_core_option_item test_env_opts[] =
 {
