+++ core/tests/unit/kmnkbd/debug_api.tests.cpp
@@ -21,8 +21,8 @@
 #include "debug_items.hpp"
 
 #include <test_assert.h>
-#include "../emscripten_filesystem.h"
-#include "../load_kmx_file.hpp"
+#include "../helpers/emscripten_filesystem.h"
+#include "../helpers/load_kmx_file.hpp"
 
 using namespace km::core::kmx;
