+++ core/tests/unit/kmx/kmx.cpp
@@ -28,8 +28,8 @@
 
 #include <test_assert.h>
 #include <test_color.h>
-#include "../emscripten_filesystem.h"
-#include "../load_kmx_file.hpp"
+#include "../helpers/emscripten_filesystem.h"
+#include "../helpers/load_kmx_file.hpp"
 
 #include "kmx_test_source.hpp"
