+++ core/tests/unit/kmx/kmx_imx.tests.cpp
@@ -13,8 +13,8 @@
 #include "../kmnkbd/action_items.hpp"
 #include <test_assert.h>
 #include <test_color.h>
-#include "../emscripten_filesystem.h"
-#include "../load_kmx_file.hpp"
+#include "../helpers/emscripten_filesystem.h"
+#include "../helpers/load_kmx_file.hpp"
 #include "utfcodec.hpp"
 
 #include <map>
