+++ core/tests/unit/kmx/kmx_key_list.tests.cpp
@@ -14,8 +14,8 @@
 
 #include <test_assert.h>
 #include <test_color.h>
-#include "../emscripten_filesystem.h"
-#include "../load_kmx_file.hpp"
+#include "../helpers/emscripten_filesystem.h"
+#include "../helpers/load_kmx_file.hpp"
 
 using namespace km::core::kmx;
