+++ core/tests/unit/meson.build
@@ -16,8 +16,8 @@ else
 endif
 
 test_util_files = [
-  meson.current_source_dir() / 'emscripten_filesystem.cpp',
-  meson.current_source_dir() / 'load_kmx_file.cpp',
+  meson.current_source_dir() / 'helpers' / 'emscripten_filesystem.cpp',
+  meson.current_source_dir() / 'helpers' / 'load_kmx_file.cpp',
 ]
 
 common_test_files = [
